Move view transaction to transaction folder
