+++ health-entity-web/src/app/components/transaction/view-transactions/view-transactions.component.html
